pyproject.toml
[project]
name = "yt-transcriber"
version = "1.0.0"
description = "YouTube Video Transcription & Summarization CLI"
readme = "README.md"
requires-python = ">=3.13"
authors = [
    { name = "Sergio Marquez", email = "sergiomarquezdev@gmail.com" }
]
keywords = ["youtube", "transcription", "whisper", "ai", "cli", "summarization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "yt-dlp>=2024.0.0",
    "openai-whisper>=20231117",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.24.0",
    "soundfile>=0.12.0",
    "tqdm>=4.65.0",
    "google-generativeai>=0.3.0",
    "rich>=13.7.0",
    "openai>=2.2.0",
    "anthropic>=0.69.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.13.0",
    "mypy>=1.18.0",
]

[project.scripts]
yt-transcriber = "yt_transcriber.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["yt_transcriber", "core"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.13"
ignore_missing_imports = true
